Fix Get and Get::Last select_statement position type

+++ lib/message_store/postgres/get/select_statement.rb
@@ -35,7 +35,7 @@ module MessageStore
             SELECT
               id::varchar,
               stream_name::varchar,
-              position::int,
+              position::bigint,
               type::varchar,
               global_position::bigint,
               data::varchar,
